requirements.txt
bcrypt
passlib
zeroc-ice
sqlalchemy
